+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+* Added Stringable support to Reference interface
+
 ## v0.1.0 (2024-01-19)
 * Ported initial Reference implementation